cook-agent/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
openai==1.3.9
google-genai==1.69.0
httpx==0.25.1
beautifulsoup4==4.12.2
lxml==4.9.3
playwright==1.40.0
playwright-stealth==1.0.1
fake-useragent==1.4.0
mcp==0.1.1
streamlit==1.28.1
requests==2.31.0
sse-starlette==2.1.2
python-dotenv==1.0.0
pytest==8.3.5
pytest-asyncio==0.25.3
